requirements.txt
streamlit==1.31.0
pandas==2.2.0
openpyxl==3.1.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.1